requirements.txt
streamlit
pandas
python-pptx
sqlalchemy
psycopg2-binary
openpyxl
